Rename `utils.gst` -> `utils.pipeline`
